make assert work for BQ
