+++ examples/sample_01/terraform/main.tf
@@ -38,32 +38,3 @@ resource "google_bigquery_table" "car" {
 EOF
 
 }
-resource "google_bigquery_table" "promoted_car" {
-  project    = var.project
-  dataset_id = google_bigquery_dataset.dataset.dataset_id
-  table_id   = "promoted_car"
-
-  schema = <<EOF
-[
-  {
-    "name": "id",
-    "type": "INTEGER",
-    "mode": "REQUIRED",
-    "description": "The license plate of the car "
-  },
-  {
-    "name": "brand",
-    "type": "STRING",
-    "mode": "NULLABLE",
-    "description": "The brand of the car "
-  },
-  {
-    "name": "new_price",
-    "type": "INTEGER",
-    "mode": "NULLABLE",
-    "description": "The new price of the car "
-  }
-]
-EOF
-
-}
