+++ examples/sample_01/test_sql/case1/assert/test_bq/jarvis_sample_01/car.json
@@ -2,6 +2,6 @@
   {
     "id": 1,
     "brand": "BMW",
-	"price": 1000
+    "price": 1000
   }
 ]
