--- src/main/java/ai/aliz/talendtestrunner/service/executor/Executor.java
@@ -1,8 +1,0 @@
-package ai.aliz.talendtestrunner.service.executor;
-
-import ai.aliz.jarvis.testconfig.ExecutionActionConfig;
-
-public interface Executor {
-    
-    public void execute(ExecutionActionConfig config);
-}
