+++ src/test/java/ai/aliz/talendtestrunner/actionConfig/AssertActionConfigTest.java
@@ -1,24 +1,24 @@
 package ai.aliz.talendtestrunner.actionConfig;
 
-import ai.aliz.talendtestrunner.service.ActionConfigForBq;
-import ai.aliz.talendtestrunner.service.AssertActionConfigCreator;
-import ai.aliz.talendtestrunner.service.AssertServiceTest;
-import ai.aliz.jarvis.testconfig.AssertActionConfig;
-
-import com.google.cloud.bigquery.BigQuery;
 import lombok.SneakyThrows;
-
-import org.junit.Ignore;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.io.File;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+
+import com.google.cloud.bigquery.BigQuery;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import ai.aliz.jarvis.testconfig.AssertActionConfig;
+import ai.aliz.talendtestrunner.service.AssertServiceTest;
+
+import org.junit.Ignore;
+import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
@@ -27,10 +27,6 @@ import static org.junit.Assert.assertThat;
 @SpringBootTest
 @Ignore
 public class AssertActionConfigTest {
-    
-    private ActionConfigForBq actionConfigForBq = new ActionConfigForBq();
-    
-    private AssertActionConfigCreator assertActionConfigCreator = new AssertActionConfigCreator();
     
     @MockBean
     private BigQuery bigQuery;
@@ -53,29 +49,29 @@ public class AssertActionConfigTest {
     @Test
     @SneakyThrows
     public void testCreateAssertActionConfig() {
-//        Map<String, Object> defaultP = new HashMap<>();
-//        ContextLoader contextLoader = new ContextLoader(new ObjectMapper());
-//        String contextPath = new File(Objects.requireNonNull(AssertServiceTest.class.getClassLoader().getResource("test_context.json").getFile())).getPath();
-//        contextLoader.parseContext(contextPath);
-//        String configPath = new File(Objects.requireNonNull(AssertServiceTest.class.getClassLoader().getResource("test_structure").getFile())).getPath() + File.separatorChar + "test_json";
-//        List<AssertActionConfig> assertActionConfigs = assertActionConfigCreator.getAssertActionConfigs(contextLoader, defaultP, new File(configPath));
-//        AssertActionConfig assertActionConfig = assertActionConfigs.get(0);
-//
-//
-//        assertThat(assertActionConfig.getProperties().get("sourcePath"), is(configPath + TestHelper.addSeparator("\\assert\\TEST_ID\\test_dataset\\assertTest.json")));
-//        assertThat(assertActionConfig.getSystem(), is("TEST_ID"));
-//        assertThat(assertActionConfig.getType(), is("AssertDataEquals"));
+        //        Map<String, Object> defaultP = new HashMap<>();
+        //        ContextLoader contextLoader = new ContextLoader(new ObjectMapper());
+        //        String contextPath = new File(Objects.requireNonNull(AssertServiceTest.class.getClassLoader().getResource("test_context.json").getFile())).getPath();
+        //        contextLoader.parseContext(contextPath);
+        //        String configPath = new File(Objects.requireNonNull(AssertServiceTest.class.getClassLoader().getResource("test_structure").getFile())).getPath() + File.separatorChar + "test_json";
+        //        List<AssertActionConfig> assertActionConfigs = assertActionConfigCreator.getAssertActionConfigs(contextLoader, defaultP, new File(configPath));
+        //        AssertActionConfig assertActionConfig = assertActionConfigs.get(0);
+        //
+        //
+        //        assertThat(assertActionConfig.getProperties().get("sourcePath"), is(configPath + TestHelper.addSeparator("\\assert\\TEST_ID\\test_dataset\\assertTest.json")));
+        //        assertThat(assertActionConfig.getSystem(), is("TEST_ID"));
+        //        assertThat(assertActionConfig.getType(), is("AssertDataEquals"));
     }
     
     private AssertActionConfig doTest(Map<String, Object> defaultP) {
         File file = new File(Objects.requireNonNull(AssertServiceTest.class.getClassLoader().getResource("test.json").getFile()));
-        AssertActionConfig assertActionConfig = actionConfigForBq.getAssertActionConfigForBq(defaultP, "testSystem", "testDataset", file);
+        //        AssertActionConfig assertActionConfig = actionConfigForBq.getAssertActionConfigForBq(defaultP, "testSystem", "testDataset", file);
+        //
+        //        assertThat(assertActionConfig.getSystem(), is("testSystem"));
+        //        assertThat(assertActionConfig.getType(), is("AssertDataEquals"));
+        //        assertThat(assertActionConfig.getProperties().get("dataset"), is("testDataset"));
+        //        assertThat(assertActionConfig.getProperties().get("sourcePath"), is(file.getPath()));
         
-        assertThat(assertActionConfig.getSystem(), is("testSystem"));
-        assertThat(assertActionConfig.getType(), is("AssertDataEquals"));
-        assertThat(assertActionConfig.getProperties().get("dataset"), is("testDataset"));
-        assertThat(assertActionConfig.getProperties().get("sourcePath"), is(file.getPath()));
-        
-        return assertActionConfig;
+        return null;
     }
 }
